feat(compiler): parse data section (#118)
* feat(zinkc): introduce parser

* ci(typos): enable typos

* feat(zinkc): split parser into methods

* feat(zinkc): parse offset and size for logs

* refactor(codegen): rename system module to call

* refactor(codegen): remove unnecessary program counters

* chore(compiler): ignore log tests

+++ codegen/src/lib.rs
@@ -44,4 +44,4 @@ pub type Imports = Vec<Func>;
 /// Data section conversion
 ///
 /// NOTE: current only support constant expression.
-pub type DataSet = BTreeMap<i64, Vec<u8>>;
+pub type DataSet = BTreeMap<i32, Vec<u8>>;
